+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Enhanced Bottons of Log Live-View
 * Git-Repository Cleanup-Hook (Post Job-Run)
 * Refactored runtime secret-handling (`ansible-runner <https://ansible.readthedocs.io/projects/runner/en/latest/intro>`_)
+* Disabling quick-execution button if custom-execution form is open
 
 ----